+++ src/app/layout.tsx
@@ -1,6 +1,6 @@
-import '/styles/globals.css';
+import '../styles/globals.css';
 import { PropsWithChildren } from 'react';
-import Navbar from 'components/Navbar';
+import Navbar from 'src/components/Navbar';
 
 export default function RootLayout({ children }: PropsWithChildren) {
   return (
