+++ src/app/users/[id]/loading.tsx
@@ -1,4 +1,4 @@
-import PlaceHolder from 'components/PlaceHolder';
+import PlaceHolder from 'src/components/PlaceHolder';
 
 export default function loading() {
   return <PlaceHolder />;
