+++ src/app/users/page.tsx
@@ -1,5 +1,5 @@
-import UsersList from 'components/UsersList';
-import { getUsers } from 'lib/api';
+import UsersList from 'src/components/UsersList';
+import { getUsers } from 'src/app/lib/api';
 
 export default async function UsersPage() {
   const users = await getUsers();
