+++ src/components/LoadingSkeleton.tsx
@@ -1,5 +1,3 @@
-import React from 'react';
-
 export default function LoadingSkeleton() {
   return (
     <section className='py-5 px-3'>
